// frontend/lib/api/types.ts
export interface components {
    schemas: {
// ...
        Invoice: {
            /** Format: uuid */
            id?: string;
            /** Format: uuid */
            subscription_id?: string | null;
            /** Format: int64 */
            amount_minor?: number;
            /** @default NGN */
            currency: string;
            /** @enum {string} */
            status?: "draft" | "open" | "paid" | "void" | "uncollectible";
            /** Format: date-time */
            period_start?: string | null;
            /** Format: date-time */
            period_end?: string | null;
            /** Format: date-time */
            issued_at?: string;
        };
// ...
    };
// ...
}